requirements.txt
numpy>=1.24.0
pandas>=2.0.0
snowflake-connector-python>=3.0.0
statsmodels>=0.14.0
matplotlib>=3.7.0
